requirements.txt
dblpy
discord.py
jishaku
emoji
git+https://github.com/dragdev-studios/guildmanager
matplotlib~=3.2.1
aiohttp
nest_asyncio
sqlalchemy
pymysql